ai_reviewer_arena/requirements.txt
gradio
PyMuPDF
Pillow
